requirements.txt
Scrapy==1.2.0
PyYAML==3.12
